# Add vzeroupper instruction for AVX512 optimization in AES-XTS