--- a/crypto/fipsmodule/aes/asm/aesni-xts-avx512.pl
+++ b/crypto/fipsmodule/aes/asm/aesni-xts-avx512.pl
@@ -2054,8 +2054,6 @@ ___
   and 	 \$0xf,$length
   je 	 .L_ret_${rndsuffix}
   jmp 	 .L_steal_cipher_next_${rndsuffix}
-  vzeroupper
-  ret
   .cfi_endproc
 ___
   }
@@ -3009,8 +3007,6 @@ ___
   add 	         \$0x10,$output
   vmovdqa 	 %xmm1,%xmm8
   jmp 	         .L_done_${rndsuffix}
-  vzeroupper
-  ret
   .cfi_endproc
 ___
   }