--- a/generated-src/linux-x86_64/crypto/fipsmodule/aesni-xts-avx512.S
+++ b/generated-src/linux-x86_64/crypto/fipsmodule/aesni-xts-avx512.S
@@ -1938,8 +1938,6 @@ aes_hw_xts_encrypt_avx512:
 	andq	$0xf,%rdx
 	je	.L_ret_hEgxyDlCngwrfFe
 	jmp	.L_steal_cipher_next_hEgxyDlCngwrfFe
-	vzeroupper
-	.byte	0xf3,0xc3
 .cfi_endproc	
 .globl	aes_hw_xts_decrypt_avx512
 .hidden aes_hw_xts_decrypt_avx512
@@ -5181,8 +5179,6 @@ aes_hw_xts_decrypt_avx512:
 	addq	$0x10,%rsi
 	vmovdqa	%xmm1,%xmm8
 	jmp	.L_done_amivrujEyduiFoi
-	vzeroupper
-	.byte	0xf3,0xc3
 .cfi_endproc	
 .section	.rodata
 .align	16