--- a/generated-src/mac-x86_64/crypto/fipsmodule/aesni-xts-avx512.S
+++ b/generated-src/mac-x86_64/crypto/fipsmodule/aesni-xts-avx512.S
@@ -1036,6 +1036,7 @@ L$_ret_hEgxyDlCngwrfFe:
 	movq	368(%rsp),%rbx
 	movq	%rbp,%rsp
 	popq	%rbp
+	vzeroupper
 	.byte	0xf3,0xc3
 
 L$_less_than_128_bytes_hEgxyDlCngwrfFe:
@@ -1937,6 +1938,7 @@ L$_num_blocks_is_1_hEgxyDlCngwrfFe:
 	andq	$0xf,%rdx
 	je	L$_ret_hEgxyDlCngwrfFe
 	jmp	L$_steal_cipher_next_hEgxyDlCngwrfFe
+	vzeroupper
 	.byte	0xf3,0xc3
 
 .globl	_aes_hw_xts_decrypt_avx512
@@ -3562,6 +3564,7 @@ L$_ret_amivrujEyduiFoi:
 	movq	368(%rsp),%rbx
 	movq	%rbp,%rsp
 	popq	%rbp
+	vzeroupper
 	.byte	0xf3,0xc3
 
 L$_less_than_128_bytes_amivrujEyduiFoi:
@@ -5178,6 +5181,7 @@ L$_done_1_amivrujEyduiFoi:
 	addq	$0x10,%rsi
 	vmovdqa	%xmm1,%xmm8
 	jmp	L$_done_amivrujEyduiFoi
+	vzeroupper
 	.byte	0xf3,0xc3
 
 .section	__DATA,__const