vzeroupper instruction would eliminate the potential
    performance penalties when switching between AVX512,
    AVX and SSE instructions.
    
    Perf data from Ice Lake -
    Did 36956000 AES-256-XTS encrypt init operations in 3000055us (12318440.8 ops/sec)
    Did 112556000 AES-256-XTS encrypt (16 bytes) operations in 3000009us (37518554.1 ops/sec): 600.3 MB/s
    Did 62446000 AES-256-XTS encrypt (256 bytes) operations in 3000011us (20815257.0 ops/sec): 5328.7 MB/s
    Did 23679000 AES-256-XTS encrypt (1350 bytes) operations in 3000079us (7892792.2 ops/sec): 10655.3 MB/s
    Did 5394000 AES-256-XTS encrypt (8192 bytes) operations in 3000100us (1797940.1 ops/sec): 14728.7 MB/s
    Did 2766000 AES-256-XTS encrypt (16384 bytes) operations in 3000348us (921893.1 ops/sec): 15104.3 MB/s
    Did 33738500 AES-256-XTS decrypt init operations in 3000021us (11246087.9 ops/sec)
    Did 100957250 AES-256-XTS decrypt (16 bytes) operations in 3000007us (33652338.1 ops/sec): 538.4 MB/s
    Did 60613500 AES-256-XTS decrypt (256 bytes) operations in 3000007us (20204452.9 ops/sec): 5172.3 MB/s
    Did 23056500 AES-256-XTS decrypt (1350 bytes) operations in 3000010us (7685474.4 ops/sec): 10375.4 MB/s
    Did 5351000 AES-256-XTS decrypt (8192 bytes) operations in 3000108us (1783602.5 ops/sec): 14611.3 MB/s
    Did 2756000 AES-256-XTS decrypt (16384 bytes) operations in 3001080us (918336.1 ops/sec): 15046.0 MB/s
    
    Perf data from SPR -
    Did 52875000 AES-256-XTS encrypt init operations in 3000011us (17624935.4 ops/sec)
    Did 165867500 AES-256-XTS encrypt (16 bytes) operations in 3000001us (55289148.2 ops/sec): 884.6 MB/s
    Did 87952000 AES-256-XTS encrypt (256 bytes) operations in 3000016us (29317177.0 ops/sec): 7505.2 MB/s
    Did 30390000 AES-256-XTS encrypt (1350 bytes) operations in 3000082us (10129723.1 ops/sec): 13675.1 MB/s
    Did 6087500 AES-256-XTS encrypt (8192 bytes) operations in 3000103us (2029097.0 ops/sec): 16622.4 MB/s
    Did 3102000 AES-256-XTS encrypt (16384 bytes) operations in 3000143us (1033950.7 ops/sec): 16940.2 MB/s
    Did 49802500 AES-256-XTS decrypt init operations in 3000013us (16600761.4 ops/sec)
    Did 146827000 AES-256-XTS decrypt (16 bytes) operations in 3000005us (48942251.8 ops/sec): 783.1 MB/s
    Did 86624500 AES-256-XTS decrypt (256 bytes) operations in 3000003us (28874804.5 ops/sec): 7391.9 MB/s
    Did 29901750 AES-256-XTS decrypt (1350 bytes) operations in 3000013us (9967206.8 ops/sec): 13455.7 MB/s
    Did 6090250 AES-256-XTS decrypt (8192 bytes) operations in 3000083us (2030027.2 ops/sec): 16630.0 MB/s
    Did 3106000 AES-256-XTS decrypt (16384 bytes) operations in 3000414us (1035190.5 ops/sec): 16960.6 MB/s