@pbadari 2023-08-01
> Where there is a conditional jump, can `vzeroupper` be placed before it (in light of zenbleedI don;t think Zenbleed